requirements.txt
streamlit
tempfile
python-dotenv
openai
pandas
sqlalchemy
plotly_express
pandasql
PyMySQL
#mysql
#mysql-connector-python

